requirements.txt
streamlit
pymupdf
pillow
openai
google-generativeai
pytesseract
opencv-python
google-genai

